imitate method applied in Anderson et al. (2017)
> To assess the performance of the PLS model, we use k-fold cross
validation [55] and a separate test-set. We divided the full set of
laboratory data into five folds, four of which are used for cross
validation and are combined as the final training set. The fifth fold is
withheld and used as a test set to validate the final model. Folds were
defined separately for each element, after removal of outliers. To
ensure that each fold represented the full elemental compositional
variation, the samples were sorted based on the major element of
interest, and samples were then assigned sequentially to each fold. In
some cases, the database contains only a small number of targets with
very high or very low compositions for a given element. These extreme
targets can have a strong influence on the model and enable it to handle
a wider range of compositions, so they are forced to be in the training
folds rather than the test set. Unless otherwise specified, the results
presented are the test set results.